fix(server): correct prod log level + consolidate nix sanitizer & JSON-body parsing
Three consolidation items from the codebase audit, bundled (net -73 LOC):

1. Fix prod log level. logger.ts read `(globalThis as any).ENVIRONMENT`,
   which is never assigned anywhere, so production silently logged at debug
   (verbose + costly). Read process.env.ENVIRONMENT (set by the Helm chart).

2. Consolidate the duplicated `nixPackageAttrRef` security sanitizer. It was
   copy-pasted byte-for-byte in the gateway orchestrator and the
   connector-worker executor, "kept in lockstep" by comment — a sandbox-escape
   drift hazard. Now a single impl in @lobu/connector-sdk (narrow ./nix-package
   subpath export, zero transitive deps) takes an error-factory param so the
   gateway keeps its OrchestratorError while the worker keeps a plain Error.
   Both existing test contracts preserved.

3. Add a `parseJsonBody` route helper and adopt it at the 7 sites that
   hand-rolled `try { await c.req.json() } catch { 400 }`. OAuth's
   invalid_request-shaped sites left as-is (different contract).

Deferred (not safe to bundle as no-behavior-change refactors):
- configsEqual/configsShallowEqual are deep vs shallow — unifying changes
  multi-replica change-detection behavior; needs its own tested change.
- generateId: the ~8 sites use divergent id formats; cosmetic, skip.
- google_calendar -> createHttpClient: changes error-message format; needs
  live-Calendar e2e per the connector gate.
